Small refactoring, now using long long ...

+++ arbre.h
@@ -35,7 +35,7 @@ extern "C" {
 #define MAX_FACTORS 30
 
 /*typedef struct t_element {
-	unsigned elt;
+	long long unsigned elt;
 } t_element;*/
 
 typedef int t_element;
@@ -44,13 +44,13 @@ typedef struct t_noeud {
 	//t_element valeur;
 	t_element cle;
 	t_element* valeur;
-	unsigned val_size;
-	int hauteur;
+	long long unsigned val_size;
+	unsigned hauteur;
 	struct t_noeud *gauche;
 	struct t_noeud *droit;
 } t_arbre;
 
-t_arbre *inserer_arbre(t_arbre *a, t_element v, t_element* factors, unsigned v_size); // Déclaration, utile pour la suite, la définitione est plus bas.
+t_arbre *inserer_arbre(t_arbre *a, t_element v, t_element* factors, long long unsigned v_size); // Déclaration, utile pour la suite, la définitione est plus bas.
 t_arbre *equilibrer_arbre(t_arbre *a);
 t_arbre *creer_arbre(t_element e, t_element* valeur, t_arbre *g, t_arbre *d);
